+++ cli/scripts/api.py
@@ -477,7 +477,7 @@ def info(p_json, p_home, p_repo, print_flag=True):
         glibc_v_display = " glibc-" + glibcV + "-"
 
     print(bold_start + ("#" * 70) + bold_end)
-    print(bold_start + "#  pgEdge CTL: " + bold_end + "v" + ver + "  " + p_home)
+    print(bold_start + "#  pgEdge CLI: " + bold_end + "v" + ver + "  " + p_home)
     print(
         bold_start
         + "# User & Host: "